Change zip button added and working
