+++ src/components/WeatherWidget.js
@@ -4,8 +4,6 @@ import ReactLoading from 'react-loading';
 import zipcodes from 'zipcodes'
 import WeatherItem from './WeatherItem'
 import { db, firebaseAuth } from '../config/constants'
-import { weatherKey } from '../keys'
-import DarkSkyApi from 'dark-sky-api';
 import ZipCodeEnter from './ZipCodeEnter'
 
 
@@ -15,9 +13,11 @@ export default class WeatherWidget extends Component {
     this.state = {
       zipCode: '',
       weather: {},
-      location: {}
+      location: {},
+      changeZip: false
     };
-    this.submitHandler = this.submitHandler.bind(this)
+    this.stateSetter = this.stateSetter.bind(this)
+    this.changeZipClickHandler = this.changeZipClickHandler.bind(this)
   }
 
   componentDidMount() {
@@ -75,80 +75,22 @@ export default class WeatherWidget extends Component {
 
   }
 
-  submitHandler(event) {
-    event.preventDefault()
-    let zip = event.target.zip.value;
+  stateSetter(newStateProps) {
+    this.setState(newStateProps)
+  }
 
-    while (zip.length < 5) zip = '0' + zip;
+  changeZipClickHandler() {
 
-    const location = zipcodes.lookup(zip)
-
-    if (!location) {
-      alert('Invalid Zipcode, try again')
-      return false
-    }
-
-    DarkSkyApi.apiKey = weatherKey;
-
-    const position = {
-      latitude: location.latitude,
-      longitude: location.longitude
-    };
-
-    DarkSkyApi.loadForecast(position)
-      .then(result => {
-        const weatherToday = result.daily.data[0]
-
-        // write zipcode to current user
-
-        // Get current users
-        const userEmail = firebaseAuth().currentUser.providerData[0].email;
-
-        // Check to see if the user has an email
-        db
-          .collection('users')
-          .doc(userEmail)
-          .set({ zip }, { merge: true })
-          .then(() => {
-            console.log('zip', zip)
-            console.log('weather', weatherToday)
-            const weather = {
-              apparentTemperatureHigh: weatherToday.apparentTemperatureHigh,
-              apparentTemperatureLow: weatherToday.apparentTemperatureLow,
-              icon: weatherToday.icon,
-              summary: weatherToday.summary
-            }
-            // save weather for this zipcode
-            db
-              .collection('weather')
-              .doc('days')
-              .collection(dateMaker())
-              .doc('zip')
-              .collection(zip)
-              .doc('forecast')
-              .set(weather)
-              .then(() => {
-                this.setState({
-                  weather,
-                  location
-                })
-              })
-              .catch(console.error)
-          })
-          .catch(console.error)
-      })
-      .catch(console.error);
-
+    this.setState({changeZip: true})
   }
 
   render() {
 
-    if (!Object.keys(this.state.location).length) {
+    if (!Object.keys(this.state.location).length || this.state.changeZip) {
       return (
-        <ZipCodeEnter submitHandler={this.submitHandler} />
+        <ZipCodeEnter stateSetter={this.stateSetter} />
       );
     }
-
     else if (Object.keys(this.state.weather).length === 0) {
       return (
         <Paper zDepth={2} className="article-card">
@@ -164,7 +106,10 @@ export default class WeatherWidget extends Component {
       <div id="wea_widget">
         <Paper zDepth={2} className="article-card">
           <div className="weather-widget">
-            <h3>Weather for today in {`${location.city}, ${location.state}`}</h3>
+            <div id="widget_header">
+              <h3>Weather for today in {`${location.city}, ${location.state}`}</h3>
+              <button id="zipButton" onClick={this.changeZipClickHandler}>Change Zip?</button>
+            </div>
             <div>
               <WeatherItem forecast={weather} />
             </div>
