+++ src/index.css
@@ -476,3 +476,15 @@ body {
   margin-left: auto;
   margin-right: auto;
 }
+
+#widget_header {
+  display: flex;
+  position: relative;
+}
+
+#zipButton {
+  position: absolute;
+  top: 0px;
+  right: 0px;
+  margin-left: 10em;
+}
